SivisoMapSchedule current location use a listener instead of last known location
Use a listener instead of last known location because the computer emulator didn't have a last known location and was crashing the project.

The listener now has a boolean that makes it do nothing after it's changed locations once and you are able to set the boolean again so that you can again move to the current location the next time the location is changed.