Refactor Permissions into own activity so far fine location
Refactoring Permissions asking into it's own activity so that I can describe what my app uses each permission for before the user has to make a decision on whether or not to accept or not.

So far I've refactored the fine location permission and need to add in the notification policy permission.